+++ CatalogService/CatalogService/Queries/Product/ProductCountQuery.cs
@@ -4,8 +4,8 @@ namespace CatalogService.Queries.Product;
 
 public class ProductCountQuery
 {
-    public ulong? MinPrice { get; set; } = null;
-    public ulong? MaxPrice { get; set; } = null;
+    public decimal? MinPrice { get; set; } = null;
+    public decimal? MaxPrice { get; set; } = null;
     public byte? MinRating { get; set; } = null;
     public byte? MaxRating { get; set; } = null;
     public string? Name { get; set; } = null;
