+++ CatalogService/CatalogService/Queries/Product/ProductGetAllQuery.cs
@@ -6,8 +6,8 @@ public class ProductGetAllQuery
 {
     public int PageNumber { get; set; } = 1;
     public int PageSize { get; set; } = 20;
-    public ulong? MinPrice { get; set; } = null;
-    public ulong? MaxPrice { get; set; } = null;
+    public decimal? MinPrice { get; set; } = null;
+    public decimal? MaxPrice { get; set; } = null;
     public byte? MinRating { get; set; } = null;
     public byte? MaxRating { get; set; } = null;
     public string? Name { get; set; } = null;
